The solr schema and config is now part of the binary distribution. This enables dependent modules to integration tests using embedded solr, and also allows us to do things like runtime analysis of the core on clients. I'm not sure about this solution however because when cores include libs (which I hope we won't need in search 2.0), velocity templates etc it becomes messy. Also I don't think solr should use these files because we have the strategy that we allow customization of field definition in cores.
Because it solves two major testing requirements - compatibility with solr test-framework and availability in dependent modules - I'll stick with it until someone comes up with a better solution.

We already do quite intricate externals definitions in https://pds-svn-gbg.pdsvision.net/svn/subversion/cms/trunk/solr/home/ so we can still add the core to admin-reference, and we because of these externals customization is not enabled yet anyway so I guess there are no immediate regressions.

Renamed the raw solr test to Core... because i think it will become a pattern for testing solr schema requirements.